Worked on the chip manager
